requirements.txt
Django<3.2 # Latest 3.1 release
python-dotenv<1 # Latest 0.x release
djangorestframework<3.13
drf-yasg<1.21
dj-database-url
django-cors-headers
django-3-jet
mysqlclient
django-colorfield
#Sphinx # Docs
#tornado # Uncomment if an http server is needed